+++ src/main/java/org/sosly/villageworks/data/zones/AABBVillageZone.java
@@ -1,19 +1,17 @@
 package org.sosly.villageworks.data.zones;
 
-import net.minecraft.core.BlockPos;
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.phys.AABB;
-import org.sosly.villageworks.api.data.IVillageZone;
-import org.sosly.villageworks.api.data.ZoneShape;
-import org.sosly.villageworks.api.data.ZoneType;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import net.minecraft.core.BlockPos;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.phys.AABB;
+import org.sosly.villageworks.api.data.ZoneShape;
+import org.sosly.villageworks.api.data.ZoneType;
 
-public class AABBVillageZone extends AbstractVillageZone implements IVillageZone {
+public class AABBVillageZone extends AbstractVillageZone {
 
     private AABB bounds;
     private List<BlockPos> cachedPOIs;
@@ -28,7 +26,7 @@ public class AABBVillageZone extends AbstractVillageZone implements IVillageZone
         super(uuid, type, name);
         this.bounds = bounds;
     }
-    
+
     public AABBVillageZone(UUID uuid, ZoneType type, int id, String name, AABB bounds, Level level) {
         super(uuid, type, id, name, level);
         this.bounds = bounds;